+++ server/api/workshop/index.js
@@ -6,6 +6,7 @@ import asyncWrapper from '../../middleware/async-wrapper'; // only wrap async fu
 const router = Router();
 
 router.get('/', asyncWrapper(controller.index));
+router.get('/active', asyncWrapper(controller.active));
 router.put('/:id', hasRole('admin'), asyncWrapper(controller.upsert)); // admin, update existing workshop
 router.delete('/:id', hasRole('admin'), asyncWrapper(controller.destroy)); // admin, delete workshop and related sections